+++ adam-apis/src/test/java/org/bdgenomics/adam/api/java/JavaADAMContigConduit.java
@@ -15,29 +15,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.apis.java;
+package org.bdgenomics.adam.api.java;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import org.bdgenomics.adam.rdd.ADAMContext;
-import org.bdgenomics.adam.rdd.variant.GenotypeRDD;
+import org.bdgenomics.adam.rdd.contig.NucleotideContigFragmentRDD;
 
 /**
- * A simple test class for the JavaADAMRDD/Context. Writes an RDD of annotations to
- * disk and reads it back.
+ * A simple test class for the JavaADAMRDD/Context. Writes an RDD of nucleotide
+ * contig fragments to disk and reads it back.
  */
-final class JavaADAMGenotypeConduit {
-    public static GenotypeRDD conduit(final GenotypeRDD recordRdd,
-                                      final ADAMContext ac) throws IOException {
+final class JavaADAMContigConduit {
+    public static NucleotideContigFragmentRDD conduit(final NucleotideContigFragmentRDD recordRdd,
+                                                      final ADAMContext ac) throws IOException {
 
         // make temp directory and save file
         Path tempDir = Files.createTempDirectory("javaAC");
-        String fileName = tempDir.toString() + "/testRdd.genotype.adam";
-        recordRdd.saveAsParquet(fileName);
+        String fileName = tempDir.toString() + "/testRdd.contig.adam";
+        recordRdd.save(fileName);
 
         // create a new adam context and load the file
         JavaADAMContext jac = new JavaADAMContext(ac);
-        return jac.loadGenotypes(fileName);
+        return jac.loadContigFragments(fileName);
     }
 }
