+++ adam-apis/src/test/java/org/bdgenomics/adam/api/java/JavaADAMFeatureConduit.java
@@ -15,29 +15,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.apis.java;
+package org.bdgenomics.adam.api.java;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import org.bdgenomics.adam.rdd.ADAMContext;
-import org.bdgenomics.adam.rdd.contig.NucleotideContigFragmentRDD;
+import org.bdgenomics.adam.rdd.feature.FeatureRDD;
 
 /**
- * A simple test class for the JavaADAMRDD/Context. Writes an RDD of nucleotide
- * contig fragments to disk and reads it back.
+ * A simple test class for the JavaADAMRDD/Context. Writes an RDD of features to
+ * disk and reads it back.
  */
-final class JavaADAMContigConduit {
-    public static NucleotideContigFragmentRDD conduit(final NucleotideContigFragmentRDD recordRdd,
-                                                      final ADAMContext ac) throws IOException {
+final class JavaADAMFeatureConduit {
+    public static FeatureRDD conduit(final FeatureRDD recordRdd,
+                                     final ADAMContext ac) throws IOException {
 
         // make temp directory and save file
         Path tempDir = Files.createTempDirectory("javaAC");
-        String fileName = tempDir.toString() + "/testRdd.contig.adam";
-        recordRdd.save(fileName);
+        String fileName = tempDir.toString() + "/testRdd.feature.adam";
+        recordRdd.save(fileName, false, false);
 
         // create a new adam context and load the file
         JavaADAMContext jac = new JavaADAMContext(ac);
-        return jac.loadContigFragments(fileName);
+        return jac.loadFeatures(fileName);
     }
 }
