+++ adam-apis/src/test/java/org/bdgenomics/adam/api/java/JavaADAMReadConduit.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.apis.java;
+package org.bdgenomics.adam.api.java;
 
 import java.io.IOException;
 import java.nio.file.Files;
