+++ adam-apis/src/test/scala/org/bdgenomics/adam/api/java/JavaADAMContextSuite.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.apis.java
+package org.bdgenomics.adam.api.java
 
 import org.bdgenomics.adam.models.ReferenceRegion
 import org.bdgenomics.adam.rdd.ADAMContext
